+++ django_formwork/widgets/__init__.py
@@ -6,7 +6,6 @@ All widgets are importable from ``django_formwork.widgets``::
 """
 
 from django_formwork.widgets.combobox import ComboBox
-from django_formwork.widgets.country_input import CountryInput
 from django_formwork.widgets.datalist import DataList
 from django_formwork.widgets.date_picker import DatePicker
 from django_formwork.widgets.file_drop_zone import FileDropZone
@@ -16,7 +15,6 @@ from django_formwork.widgets.input_number import InputNumber
 from django_formwork.widgets.multi_select import MultiSelect
 from django_formwork.widgets.otp_input import OTPInput
 from django_formwork.widgets.password_reveal import PasswordReveal
-from django_formwork.widgets.phone_input import PhoneInput
 from django_formwork.widgets.range import Range
 from django_formwork.widgets.rating import Rating
 from django_formwork.widgets.search_select import SearchSelect
@@ -25,7 +23,6 @@ from django_formwork.widgets.validated_textarea import ValidatedTextarea
 
 __all__ = [
     "ComboBox",
-    "CountryInput",
     "DataList",
     "DatePicker",
     "FileDropZone",
@@ -35,7 +32,6 @@ __all__ = [
     "MultiSelect",
     "OTPInput",
     "PasswordReveal",
-    "PhoneInput",
     "Range",
     "Rating",
     "SearchSelect",